+++ Patches/Inventory/InventoryUtilitiesServerPatches.cs
@@ -1,9 +1,12 @@
+using System;
+using System.Collections.Generic;
 using Stunlock.Core;
 using VAMP;
 using HarmonyLib;
 using ProjectM;
 using Unity.Entities;
 using Unity.Collections;
+using VAMP.Data;
 using VComforts.Patches.Connection;
 using VComforts.Systems;
 
@@ -15,13 +18,107 @@ public static class TrySortInventoryPatch
     static bool Prefix(ref NativeParallelHashMap<PrefabGUID, ItemData> itemDataMap, Entity target)
     {
         if (target == Entity.Null) return true;
-        if (!target.IsPlayer()) return true;
+        if (target.IsPlayer())
+        {
+#if DEBUG
+            Plugin.LogInstance.LogWarning(
+                "TrySortInventoryPatch: Owner of inventory is a player, applying custom itemDataMap");
+#endif
+            var originalMap = Core.SystemService.GameDataSystem._GameDatas.ItemHashLookupMap;
+            itemDataMap = InitializePlayer_Patch.GetUpdatedItemDataMap(target, originalMap);
+            BonusSystem.HandleInventoryBuffs(target);
+            return true;
+        }
 
-        Plugin.LogInstance.LogWarning("Owner of inventory is a player, applying custom itemDataMap");
-        var originalMap = Core.SystemService.GameDataSystem._GameDatas.ItemHashLookupMap;
-        itemDataMap = InitializePlayer_Patch.GetUpdatedItemDataMap(target, originalMap);
-        BonusSystem.HandleInventoryBuffs(target);
         return true;
+    }
+
+    static void Postfix(Entity target)
+{
+    if (target == Entity.Null) return;
+    if (!IsBloodPotionStorage(target)) return;
+
+    var attachedBuffer = target.ReadBuffer<AttachedBuffer>();
+    var invBuffer = attachedBuffer[0].Entity.ReadBuffer<InventoryBuffer>();
+
+    var potionSlots = new List<int>();
+    var potions = new List<(InventoryBuffer item, string primaryType, float quality, bool hasSecondary, string secondaryType, float secondaryQuality, int originalIndex)>();
+
+    for (int i = 0; i < invBuffer.Length; i++)
+    {
+        var itemType = invBuffer[i].ItemType;
+        if (itemType == Prefabs.Item_Consumable_PrisonPotion_Bloodwine ||
+            itemType == Prefabs.Item_Consumable_PrisonPotion ||
+            itemType == Prefabs.Item_Consumable_PrisonPotion_Mixed)
+        {
+            var itemEntity = invBuffer[i].ItemEntity.GetEntityOnServer();
+            if (!itemEntity.Has<StoredBlood>()) continue;
+            var storedBlood = itemEntity.Read<StoredBlood>();
+            var primaryType = storedBlood.PrimaryBloodType.GetNamePrefab().Split("_")[1];
+            var quality = storedBlood.BloodQuality;
+            var hasSecondary = !storedBlood.SecondaryBlood.IsEmpty;
+            potionSlots.Add(i);
+            var secondaryType = storedBlood.SecondaryBlood.IsEmpty ? "" : storedBlood.SecondaryBlood.Type.GetNamePrefab().Split("_")[1];
+            var secondaryQuality = storedBlood.SecondaryBlood.IsEmpty ? 0f : storedBlood.SecondaryBlood.Quality;
+            potions.Add((invBuffer[i], primaryType, quality, hasSecondary, secondaryType, secondaryQuality, i));
+        }
+    }
+
+    var sortMode = Settings.CUSTOM_BLOODPOTION_SORTING_PRIMARYTHENQUALITY.Value
+        ? BloodPotionSortMode.PrimaryTypeThenQuality
+        : BloodPotionSortMode.PrimaryOnlyFirstThenSecondary;
+
+    switch (sortMode)
+    {
+        case BloodPotionSortMode.PrimaryTypeThenQuality:
+            potions.Sort((a, b) =>
+            {
+                int cmp = string.Compare(a.primaryType, b.primaryType, StringComparison.Ordinal);
+                if (cmp != 0) return cmp;
+                cmp = string.Compare(a.secondaryType, b.secondaryType, StringComparison.Ordinal);
+                if (cmp != 0) return cmp;
+                cmp = b.quality.CompareTo(a.quality);
+                if (cmp != 0) return cmp;
+                cmp = b.secondaryQuality.CompareTo(a.secondaryQuality);
+                if (cmp != 0) return cmp;
+                return a.originalIndex.CompareTo(b.originalIndex);
+            });
+            break;
+        case BloodPotionSortMode.PrimaryOnlyFirstThenSecondary:
+            potions.Sort((a, b) =>
+            {
+                if (a.hasSecondary != b.hasSecondary)
+                    return a.hasSecondary ? 1 : -1;
+                int cmp = string.Compare(a.primaryType, b.primaryType, StringComparison.Ordinal);
+                if (cmp != 0) return cmp;
+                cmp = string.Compare(a.secondaryType, b.secondaryType, StringComparison.Ordinal);
+                if (cmp != 0) return cmp;
+                cmp = b.quality.CompareTo(a.quality);
+                if (cmp != 0) return cmp;
+                cmp = b.secondaryQuality.CompareTo(a.secondaryQuality);
+                if (cmp != 0) return cmp;
+                return a.originalIndex.CompareTo(b.originalIndex);
+            });
+            break;
+        default:
+            throw new ArgumentOutOfRangeException();
+    }
+
+    for (int i = 0; i < potionSlots.Count; i++)
+    {
+        invBuffer[potionSlots[i]] = potions[i].item;
+    }
+}
+    
+    private static bool IsBloodPotionStorage(Entity entity)
+    {
+        return entity.GetPrefabGuid() == Prefabs.TM_Castle_Container_Specialized_Consumable_T01 || entity.GetPrefabGuid() == Prefabs.TM_Castle_Container_Specialized_Consumable_T02 || entity.GetPrefabGuid() == Prefabs.TM_Castle_Container_Bookshelf_6x2_Blood01 || entity.GetPrefabGuid() == Prefabs.TM_Castle_Container_Bookshelf_3x2_Blood01;
+    }
+    
+    private enum BloodPotionSortMode
+    {
+        PrimaryTypeThenQuality,
+        PrimaryOnlyFirstThenSecondary
     }
 }
 
@@ -140,6 +237,7 @@ public static class InternalTryAddItemInInventoryPatch
         {
             return true;
         }
+
         var player = inventoryEntity.Read<InventoryConnection>().InventoryOwner;
 
         if (!player.IsPlayer())